Remove deprecated functionality
